Deler opp etter package by feature for å gi bedre cohesion
