+++ src/main/kotlin/no/nav/personbruker/tms/personalia/api/config/ApplicationContext.kt
@@ -2,8 +2,9 @@ package no.nav.personbruker.tms.personalia.api.config
 
 import com.expediagroup.graphql.client.ktor.GraphQLKtorClient
 import no.nav.personbruker.tms.personalia.api.health.HealthService
-import no.nav.personbruker.tms.personalia.api.personalia.PersonaliaConsumer
-import no.nav.personbruker.tms.personalia.api.personalia.PersonaliaService
+import no.nav.personbruker.tms.personalia.api.ident.IdentService
+import no.nav.personbruker.tms.personalia.api.navn.NavnConsumer
+import no.nav.personbruker.tms.personalia.api.navn.NavnService
 import no.nav.tms.token.support.tokendings.exchange.TokendingsServiceBuilder
 import java.net.URL
 
@@ -16,8 +17,9 @@ class ApplicationContext {
     val tokendingsService = TokendingsServiceBuilder.buildTokendingsService(maxCachedEntries = 10000)
     val tokendingsTokenFetcher = TokendingsTokenFetcher(tokendingsService, environment.pdlClientId)
 
-    val personalieConsumer = PersonaliaConsumer(GraphQLKtorClient(URL(environment.pdlUrl), httpClient), environment.pdlUrl)
+    val navnConsumer = NavnConsumer(GraphQLKtorClient(URL(environment.pdlUrl), httpClient), environment.pdlUrl)
+    val navnService = NavnService(navnConsumer, tokendingsTokenFetcher)
 
-    val personaliaService = PersonaliaService(personalieConsumer, tokendingsTokenFetcher)
+    val identService = IdentService()
 
 }
