+++ src/main/kotlin/no/nav/personbruker/tms/personalia/api/ident/Ident.kt
@@ -1,4 +1,4 @@
-package no.nav.personbruker.tms.personalia.api.personalia
+package no.nav.personbruker.tms.personalia.api.ident
 
 import kotlinx.serialization.Serializable
